Fixing the scope handing of the karaf-maven-plugin
